+++ Products.API/Interfaces/IProductsService.cs
@@ -8,9 +8,9 @@ namespace Products.API.Interfaces
     public interface IProductsService
     {
         Task<IEnumerable<ProductsEntity>> ListProductsAsync(GetRequestFilter filters);
-        Task<ProductsEntity> FindProductAsync(int pId);
+        Task<ProductsEntity> FindProductAsync(long pId);
         Task<ProductsEntity> CreateProductAsync(ProductsEntity item);
-        Task<ProductsEntity> PutProductAsync(int pId, ProductsEntity item);
-        Task<bool> DeleteProductAsync(int pId);
+        Task<ProductsEntity> PutProductAsync(long pId, ProductsEntity item);
+        Task<bool> DeleteProductAsync(long pId);
     }
 }
